Complete overhaul of the package
Will now be managed as a monorepo with lerna. The original package becomes a theme of the new react-vitae package.

The entire package is now written in TypeScript and the main `react-vitae` package is thoroughly tested with Jest.

The package is now truly ready for production use, but the themes still need to be redesigned.